Add files via upload
Using a data table showing the calories, carbs, and fat content of Starbucks drinks, this R project creates a linear regression model using the lm() function. Two models are presented and analyzed for effectiveness in being able to predict the dependant variable of Calories while only knowing the independent variables of the carbohydrate and fat content of each drink.